Move to shuffle/tests and add an assert.
Move to shuffle/tests, implement some cleanup based on test_concurrent,
and assert that the final data returned by dask DataFrame matches the
original pandas DataFrame.
